+++ lib/src/Styled.dart
@@ -1,6 +1,17 @@
 part of '../styled_widget.dart';
 
 class Styled {
+  static Widget builder({
+    required Widget Function(BuildContext context, Widget child) builder,
+    required Widget child,
+  }) =>
+      Builder(
+        builder: (context) => builder(
+          context,
+          child,
+        ),
+      );
+
   static Widget widget({Widget? child}) =>
       child ??
       LimitedBox(